% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{bayes_glm_model}
\alias{bayes_glm_model}
\title{Sampled Bayesian GLM (class)}
\usage{
bayes_glm_model(ptr, terms, xlevels)
}
\arguments{
\item{ptr, terms, xlevels}{Internal.}
}
\value{
A \code{bayes_glm_model} object.
}
\description{
Returned by \code{\link[=bayes_glm_fit]{bayes_glm_fit()}}.
}
