% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{distortion}
\alias{distortion}
\title{Distortion risk measure}
\usage{
distortion(
  kind = c("tvar", "wang", "proportional_hazard", "dual_power"),
  param
)
}
\arguments{
\item{kind}{One of \code{"tvar"}, \code{"wang"}, \code{"proportional_hazard"} and
\code{"dual_power"}.}

\item{param}{The distortion's parameter.}
}
\value{
A \code{distortion} object with \code{kind} and \code{param} properties. Use it
with \code{\link[=risk_measure]{risk_measure()}} and \code{\link[=allocate]{allocate()}}.
}
\description{
A concave distortion \code{g} of the survival function, giving the coherent
risk measure \verb{rho(X) = integral of g(S(x)) dx}. Every kind has a parameter
value that gives the mean.
}
\details{
\tabular{lll}{
   \code{kind} \tab \code{g(s)} \tab \code{param} \cr
   \code{"tvar"} \tab \code{min(s / (1 - p), 1)} \tab \code{p} in \verb{[0, 1]} \cr
   \code{"wang"} \tab \code{pnorm(qnorm(s) + lambda)} \tab \code{lambda >= 0} \cr
   \code{"proportional_hazard"} \tab \code{s^rho} \tab \code{rho} in \verb{(0, 1]} \cr
   \code{"dual_power"} \tab \code{1 - (1 - s)^beta} \tab \code{beta >= 1} \cr
}
}
\examples{
d <- distortion("wang", 0.5)
risk_measure(sampled(c(10, 20, 30, 40, 50)), d)
distortion_g(distortion("dual_power", 2), 0.3)
}
