% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{inuring_tower}
\alias{inuring_tower}
\title{Reinsurance tower with inuring order}
\usage{
inuring_tower(stages)
}
\arguments{
\item{stages}{A list of stages, each a list of \link{xol_layer} objects. Layer
names must be unique across stages.}
}
\value{
A \link{reinsurance_tower}.
}
\description{
Stages apply in order: each stage's layers see the losses net of all
earlier stages, event by event, with annual terms used up in event order
(see \code{\link[=ceded_by_event]{ceded_by_event()}}).
}
\examples{
# A 50\% quota share inures to the benefit of a 5 xs 5 cover.
tw <- inuring_tower(list(list(quota_share("QS", 0.5)), list(xol_layer("5x5", 5, 5))))
tw@stages
tower_ceded(tw, 30)
}
