% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{variance}
\alias{variance}
\title{Variance of a distribution}
\usage{
variance(dist, ...)
}
\arguments{
\item{dist}{A \link{distribution}.}

\item{...}{Unused; for methods.}
}
\value{
A single number.
}
\description{
Variance of a distribution
}
\examples{
variance(lognormal_from_mean_cv(1000, 0.5))
}
