% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{reinsurance_tower}
\alias{reinsurance_tower}
\title{Reinsurance tower}
\usage{
reinsurance_tower(layers)
}
\arguments{
\item{layers}{A list of \link{xol_layer} objects with unique names.}
}
\value{
A \code{reinsurance_tower} object; apply it with \code{\link[=apply_tower]{apply_tower()}}.
}
\description{
Layers applied to the same ground-up losses (no inuring order yet).
}
\examples{
tw <- reinsurance_tower(list(xol_layer("5x5", 5e6, 5e6), xol_layer("15x10", 15e6, 10e6)))
tw@layer_names
}
